make platform-specific executables compile everywhere
Some Rust tools get hung up if the entire workspace cannot be compiled.

refs https://github.com/rust-lang/cargo/issues/5220
