+++ windows-redirector/build.rs
@@ -1,3 +1,4 @@
+#[cfg(windows)]
 fn main() {
     let mut res = winres::WindowsResource::new();
     res.set_manifest(
@@ -16,3 +17,6 @@ fn main() {
     res.set_icon("mitmproxy.ico");
     res.compile().unwrap();
 }
+
+#[cfg(not(windows))]
+fn main() {}
